feat:添加postgres支持

+++ genal-chat-server/src/modules/friend/entity/friendMessage.entity.ts
@@ -17,6 +17,6 @@ export class FriendMessage {
   @Column()
   messageType: string;
 
-  @Column('double')
+  @Column('int8')
   time: number;
 }
